MovieScreen

+++ app/data/src/commonMain/kotlin/ru/slartus/moca/data/api/tmdb/models/Movie.kt
@@ -20,6 +20,6 @@ data class Movie(
     val release_date: String?,
     val title: String?,
     val video: Boolean?,
-    val vote_average: Double?,
+    val vote_average: Float?,
     val vote_count: Int?
 )
